Refactor text handling to be less repetitive
The actual text decoders are still huge switch trees, but now they
utilize a special variant-like class to not be so repetitive in the text
representation aspect. This variant-like class setup also makes it much
easier to implement other renderings of the textual info, and someday
can work as an intermediate form for editing as well.