Reinstate a much-improved HexViewer
Now instead of using a couple of table views, we just draw it for
ourselves in an abstract scroll area. Note that using the widget rawly
as a window isn't permanent; it's expected we'll need an enclosing
window for various functions to perform on the data (e.g. searching).

Also, we don't yet draw a "text cursor", so there's not even a chance to
copy things or anything.