Reimplement using text viewer.
Notice we got rid of the text dock widget in the process, since it seems
unhelpful at this point. Also added a rainbow background once you start
selecting text in the viewer, so you can see the messages against a
variety of backgrounds (esp. those types that have no box to be in).
Definitely something configurable in the future.