Bump version to 0.1.0
The text viewer is at a reasonable enough point that a minor version
bump is worthwhile, in my opinion. We're done now with the big overhaul
of the interface (switching to a MDI style), and a decent-looking hex
viewer. Not the end of even the text viewer of course, just seems worth
a version bump.