Add README and make compiler options more portable
This adds the ever-so-standard README for friendliness, and makes sure
we stand a better chance at handling different compilers at the outset.
Note that this support doesn't currently take into account build types;
that is, we only have it set up for the "any build type" flags. The
build-type-specific shouldn't be hard, though.