Implement fixed number support for RCP stuff.
This uses the gmp library (which CMake of course searches for and gets)
to implement fixed numbers in a way that won't make me tear my hair out
right away. An alternate implementation without requiring gmp is
expected for the future, though it may be limited to 64bit or 32bit
numbers, depending on if your compiler supports a 128bit type, or I
decide on doing things the harder way.