Add spacer to dock widget, fix slight bug in text conversion.
First up, a space item was added to the ROM Info dock widget, so if you
decide to move around the dock widget to be on the other side, this
widget won't have everything stretched out ugily(?).

Second, get rid of the part where we add the "end message" command to
the AST, since that's now represented by the rest of the AST structure.